+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 Version latest (0.1.0, 2019-09-03)
 
-- FEAT: add setup.py and use click and now executable is just `isitfit`
 - FEAT: first release of working version
+- FEAT: add setup.py and use click and now executable is just `isitfit`
+- FEAT: add `--debug` flag